Revert "Split game and board classes"
This reverts commit 1266927cd8226e02b23fcb5f26202bf127902fdb.
